Removed deprecated PE class. Updated Mimikatz loader to use ManualMap
