+++ SharpSploit/Execution/ManualMap/Overload.cs
@@ -116,7 +116,7 @@ namespace SharpSploit.Execution.ManualMap
             }
 
             // Map decoy from disk
-            Execute.PE.PE_MANUAL_MAP DecoyMetaData = Map.MapModuleFromDisk(DecoyModulePath);
+            PE.PE_MANUAL_MAP DecoyMetaData = Map.MapModuleFromDisk(DecoyModulePath);
             IntPtr RegionSize = DecoyMetaData.PEINFO.Is32Bit ? (IntPtr)DecoyMetaData.PEINFO.OptHeader32.SizeOfImage : (IntPtr)DecoyMetaData.PEINFO.OptHeader64.SizeOfImage;
 
             // Change permissions to RW
